% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsy}
\alias{nlsy}
\title{National Longitudinal Survey of Youth, 1979 cohort}
\format{
A data frame with 5583 rows and 38 variables.
}
\usage{
nlsy
}
\description{
A dataset containing information on education, demographics, unemployment, earnings,
and mental health from the National Longitudinal Survey of Youth, 1979 cohort.
}
\keyword{datasets}
