% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwrmed.R
\name{rwrmed}
\alias{rwrmed}
\title{Regression-with-Residuals Analysis of Causal Mediation}
\usage{
rwrmed(
  treatment,
  pre_cov,
  zmodels,
  y_form,
  m_form,
  m_family = gaussian,
  weights,
  data
)
}
\arguments{
\item{treatment}{A character string indicating the name of the treatment variable, which
can be either binary or continuous.}

\item{pre_cov}{A character string vector indicating the names of the pretreatment covariates
that are to be centered at their means. Thus factor-valued covariates should be pre-coded as
dummy variables.}

\item{zmodels}{A list of fitted \code{lm} or \code{glm} objects for the
post-treatment confounders of the mediator-outcome relationship.}

\item{y_form}{Formula for the outcome model.}

\item{m_form}{Formula for the mediator model.}

\item{m_family}{The family of the mediator model to be specified in \code{\link[stats]{glm}}.}

\item{weights}{An optional vector of weights to be used in fitting the outcome and mediator models.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class \code{rwrmed}.
\item{y_model}{The fitted outcome model.}
\item{m_model}{The fitted mediator model.}
\item{z_models}{Fitted \code{lm} or \code{glm} models for the post-treatment confounders.}
\item{var_names}{Names of the treatment variable, the mediator, the outcome, and
the pretreatment and posttreatment covariates}
\item{data}{The original data frame.}
\item{data_ed}{The data frame with residualized pre- and post-treatment covariates}
\item{call}{The matched call.}
}
\description{
\code{rwrmed} is a function that implements the regression-with-residuals (RWR)
approach to causal mediation, allowing for post-treatment confounding of the mediator-outcome
relationship. Specifically, it fits user-specified mediator and outcome models with a set of
residualized post-treatment confounders. It returns an object of class \code{rwrmed}, which can be
used for effect decomposition via the \code{\link{decomp}} function.
}
\seealso{
\code{\link{decomp}} for effect decomposition based on a fitted \code{rwrmed}
object.
}
