% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsim.R
\name{medsim}
\alias{medsim}
\title{Simulate Mediated Effects with Bootstrapping}
\usage{
medsim(
  data,
  num_sim = 2000,
  cat_list = c("0", "1"),
  treatment,
  intv_med = NULL,
  model_spec,
  weights = NULL,
  seed = NULL,
  boot = FALSE,
  boot_reps = 100,
  boot_cores = NULL,
  boot_conf_level = 0.95
)
}
\arguments{
\item{data}{Data frame containing the variables.}

\item{num_sim}{Number of simulations to run. Default is \code{2000}.}

\item{cat_list}{Vector of levels for the treatment. Default is \code{c("0", "1")}.}

\item{treatment}{Name of the treatment variable.}

\item{intv_med}{Intervened mediator. Default is \code{NULL}.}

\item{model_spec}{List of model specifications, where each model specification includes the \code{func}, \code{formula}, and \code{args}.}

\item{weights}{Optional. A string specifying the name of the column in the data that contains the weights to be used in weighted regression models and for calculating weighted means. Default is \code{NULL}.}

\item{seed}{Seed for reproducibility. Default is \code{NULL}.}

\item{boot}{Logical indicating whether to perform bootstrapping. Default is \code{FALSE}.}

\item{boot_reps}{Number of bootstrap replications. Default is 100.}

\item{boot_cores}{Number of CPU cores for parallel bootstrap. Defaults to available cores minus 2.}

\item{boot_conf_level}{Confidence level for bootstrap intervals. Default is \code{0.95}.}
}
\value{
A data frame of point estimates, confidence intervals, and p-values if bootstrapping is performed; otherwise a list of point estimates
}
\description{
This function simulates mediated effects and optionally performs bootstrapping to estimate confidence intervals and p-values.
}
